+++ testng/src/main/java/cucumber/api/testng/CucumberFeatureWrapper.java
@@ -1,13 +1,11 @@
 package cucumber.api.testng;
 
-import cucumber.runtime.model.CucumberFeature;
-
 /**
  * The only purpose of this interface is to be able to provide a custom
  * <pre>toString()</pre>, making TestNG reports look more descriptive.
  *
- * @see CucumberFeatureWrapperImpl
+ * @see AbstractTestNGCucumberTests#runScenario(cucumber.api.testng.PickleEventWrapper, cucumber.api.testng.CucumberFeatureWrapper)
  */
 public interface CucumberFeatureWrapper {
-    CucumberFeature getCucumberFeature();
+
 }